external validation and refactoring
