+++ src/main/java/com/intuit/graphql/orchestrator/batch/GraphQLServiceBatchLoader.java
@@ -114,7 +114,7 @@ public class GraphQLServiceBatchLoader implements BatchLoader<DataFetchingEnviro
       MergedField filteredRootField = result.getMergedField();
       if (filteredRootField != null) {
         filteredRootField.getFields().stream()
-            .map(field -> (serviceMetadata.hasFieldResolverDirective() || serviceMetadata.isFederationService())
+            .map(field -> (serviceMetadata.shouldRemoveExternalFields())
                         ? removeFieldsWithExternalTypes(field, getRootFieldDefinition(key.getExecutionStepInfo()).getType())
                         : field)
             .forEach(selectionSetBuilder::selection);
