+++ src/main/java/com/intuit/graphql/orchestrator/schema/ServiceMetadata.java
@@ -33,4 +33,6 @@ public interface ServiceMetadata {
   boolean isEntity(String typename);
 
   FederationMetadata getFederationServiceMetadata();
+
+  boolean shouldRemoveExternalFields();
 }
