+++ src/main/java/com/intuit/graphql/orchestrator/schema/type/conflict/resolver/XtextTypeConflictResolver.java
@@ -1,18 +1,17 @@
 package com.intuit.graphql.orchestrator.schema.type.conflict.resolver;
 
+import static com.intuit.graphql.orchestrator.utils.FederationConstants.FEDERATION_EXTENDS_DIRECTIVE;
 import static com.intuit.graphql.orchestrator.utils.XtextTypeUtils.checkFieldsCompatibility;
 import static com.intuit.graphql.orchestrator.utils.XtextTypeUtils.isEntity;
 import static com.intuit.graphql.orchestrator.utils.XtextTypeUtils.isScalarType;
 import static com.intuit.graphql.orchestrator.utils.XtextTypeUtils.toDescriptiveString;
-import static com.intuit.graphql.orchestrator.utils.XtextTypeUtils.typeContainsDirective;
+import static com.intuit.graphql.orchestrator.utils.XtextUtils.definitionContainsDirective;
 
 import com.intuit.graphql.graphQL.EnumTypeDefinition;
 import com.intuit.graphql.graphQL.InterfaceTypeDefinition;
 import com.intuit.graphql.graphQL.ObjectTypeDefinition;
-import com.intuit.graphql.graphQL.ScalarTypeDefinition;
 import com.intuit.graphql.graphQL.TypeDefinition;
 import com.intuit.graphql.graphQL.UnionTypeDefinition;
-import com.intuit.graphql.graphQL.impl.ScalarTypeDefinitionImpl;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.Arrays;
@@ -50,7 +49,7 @@ public class XtextTypeConflictResolver {
       boolean conflictingTypeisEntity = isEntity(conflictingType);
       boolean existingTypeIsEntity = isEntity(existingType);
       boolean entityComparison =  conflictingTypeisEntity && existingTypeIsEntity;
-      boolean baseExtensionComparison = typeContainsDirective(existingType, "extends") || typeContainsDirective(conflictingType, "extends");
+      boolean baseExtensionComparison = definitionContainsDirective(existingType, FEDERATION_EXTENDS_DIRECTIVE) || definitionContainsDirective(conflictingType, FEDERATION_EXTENDS_DIRECTIVE);
 
       if(isEntity(conflictingType) != isEntity(existingType)) {
         throw new TypeConflictException("Type %s is conflicting with existing type %s. Only one of the types are an entity.");
