+++ src/main/java/com/intuit/graphql/orchestrator/utils/GraphQLUtil.java
@@ -90,33 +90,4 @@ public class GraphQLUtil {
     throw new CreateTypeException(
         String.format(ERR_CREATE_TYPE_UNEXPECTED_TYPE, GraphQLTypeUtil.simplePrint(graphQLType)));
   }
-
-  public static String getUniqueIdFromFieldSet(String fieldSet) {
-    if(!fieldSet.startsWith("{")) {
-      fieldSet = StringUtils.join(StringUtils.SPACE, "{", fieldSet, "}");
-    }
-
-    Document graphqlDocument = Parser.parse(fieldSet);
-
-    List<OperationDefinition> definitions = graphqlDocument.getDefinitions().stream()
-            .map(OperationDefinition.class::cast).collect(Collectors.toList());
-
-    return convertSelectionSetToUniqueId(definitions.get(0).getSelectionSet());
-  }
-
-  private static String convertSelectionSetToUniqueId(SelectionSet selectionSet) {
-    Comparator<Field> fieldComparable = Comparator.comparing(Field::getName);
-
-    List<Field> fields = selectionSet.getSelections().stream()
-            .map(Field.class::cast).sorted(fieldComparable).collect(Collectors.toList());
-
-    String directChildrenUniqueId = fields.stream().map(Field::getName).reduce("", (partialId, fieldName) -> partialId + fieldName);
-    String descendantsUniqueId = fields.stream().filter(field -> CollectionUtils.isNotEmpty(field.getChildren())).map(field -> {
-      SelectionSet childSelections =  field.getChildren().stream().map(SelectionSet.class::cast).collect(Collectors.toList()).get(0);
-      return convertSelectionSetToUniqueId(childSelections);
-    }).reduce("", (partialId, childSelectionNames) -> partialId + childSelectionNames);
-
-    return directChildrenUniqueId + descendantsUniqueId;
-  }
-
 }
