+++ src/test/java/com/intuit/graphql/orchestrator/schema/transform/FederationTransformerPostMergeTest.java
@@ -1,7 +1,10 @@
 package com.intuit.graphql.orchestrator.schema.transform;
 
+import static com.intuit.graphql.orchestrator.XtextObjectCreationUtil.buildDirective;
+import static com.intuit.graphql.orchestrator.XtextObjectCreationUtil.buildDirectiveDefinition;
 import static com.intuit.graphql.orchestrator.XtextObjectCreationUtil.buildFieldDefinition;
 import static com.intuit.graphql.orchestrator.XtextObjectCreationUtil.buildObjectTypeDefinition;
+import static com.intuit.graphql.orchestrator.utils.XtextTypeUtils.getFieldDefinitions;
 import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.times;
@@ -9,20 +12,34 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.google.common.collect.ImmutableMap;
+import com.intuit.graphql.graphQL.Argument;
+import com.intuit.graphql.graphQL.Directive;
+import com.intuit.graphql.graphQL.DirectiveDefinition;
 import com.intuit.graphql.graphQL.FieldDefinition;
 import com.intuit.graphql.graphQL.ObjectTypeDefinition;
 import com.intuit.graphql.graphQL.TypeDefinition;
+import com.intuit.graphql.graphQL.ValueWithVariable;
+import com.intuit.graphql.graphQL.impl.ArgumentImpl;
+import com.intuit.graphql.orchestrator.schema.type.conflict.resolver.TypeConflictException;
+import com.intuit.graphql.orchestrator.utils.FederationConstants;
 import com.intuit.graphql.orchestrator.xtext.XtextGraph;
+
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
+
+import org.assertj.core.api.AssertionsForClassTypes;
+import org.assertj.core.api.AssertionsForInterfaceTypes;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
 
 @RunWith(MockitoJUnitRunner.class)
-public class KeyTransformerPostMergeTest {
+public class FederationTransformerPostMergeTest {
 
   private static final FieldDefinition TEST_FIELD_DEFINITION_1 = buildFieldDefinition("testField1");
   private static final FieldDefinition TEST_FIELD_DEFINITION_2 = buildFieldDefinition("testField2");
@@ -30,7 +47,7 @@ public class KeyTransformerPostMergeTest {
   @Mock
   private XtextGraph xtextGraphMock;
 
-  private final KeyTransformerPostMerge subjectUnderTest = new KeyTransformerPostMerge();
+  private final FederationTransformerPostMerge subjectUnderTest = new FederationTransformerPostMerge();
 
   @Before
   public void setup() {
@@ -56,5 +73,4 @@ public class KeyTransformerPostMergeTest {
     verify(xtextGraphMock, times(1)).getEntitiesByTypeName();
     verify(xtextGraphMock, times(2)).getEntityExtensionsByNamespace();
   }
-
 }
